Producers/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
kafka-python==2.0.2
avro-python3==1.10.2
apache-airflow==2.8.0